refactor: remove test requires runtime
