+++ extension/Cargo.toml
@@ -15,21 +15,16 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 codec.workspace = true
 log.workspace = true
-scale-info.workspace = true
 
 # Local
-pallet-api.workspace = true
 pop-primitives.workspace = true
 pop-runtime-common.workspace = true
 
 # Substrate
 frame-support.workspace = true
 frame-system.workspace = true
-pallet-assets.workspace = true
-pallet-balances.workspace = true
 pallet-contracts.workspace = true
 sp-core.workspace = true
-sp-io.workspace = true
 sp-runtime.workspace = true
 sp-std.workspace = true
 
@@ -39,25 +34,19 @@ rand = "0.8.5"
 [features]
 default = ["std"]
 std = [
+ "log/std",
  "codec/std",
  "frame-support/std",
  "frame-system/std",
- "log/std",
- "pallet-assets/std",
- "pallet-balances/std",
  "pallet-contracts/std",
  "pop-primitives/std",
- "scale-info/std",
- "sp-core/std",
- "sp-io/std",
  "sp-runtime/std",
+ "sp-core/std",
  "sp-std/std",
 ]
 runtime-benchmarks = [
  "frame-support/runtime-benchmarks",
  "frame-system/runtime-benchmarks",
- "pallet-assets/runtime-benchmarks",
- "pallet-balances/runtime-benchmarks",
  "pallet-contracts/runtime-benchmarks",
  "sp-runtime/runtime-benchmarks",
 ]
